requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pywebview>=5.0.0
ttkbootstrap>=1.10.0
rich>=13.0.0
